feat(model, http): Make positions possible to create.
This PR moves the Position struct to the model crate.

It also fixes an issue where it was not possible to create null values
in the fields that had the ability to be null or left out.
