+++ twilight-http/src/request/guild/update_guild_channel_positions.rs
@@ -5,41 +5,15 @@ use crate::{
     response::{marker::EmptyBody, Response, ResponseFuture},
     routing::Route,
 };
-use serde::Serialize;
 use std::future::IntoFuture;
-use twilight_model::id::{
-    marker::{ChannelMarker, GuildMarker},
-    Id,
+use twilight_model::{
+    http::channel_position::Position,
+    id::{marker::GuildMarker, Id},
 };
-
-#[derive(Serialize)]
-pub struct Position {
-    id: Id<ChannelMarker>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    lock_permissions: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    parent_id: Option<Id<ChannelMarker>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    position: Option<u64>,
-}
-
-impl From<(Id<ChannelMarker>, u64)> for Position {
-    fn from((id, position): (Id<ChannelMarker>, u64)) -> Self {
-        Self {
-            id,
-            lock_permissions: None,
-            parent_id: None,
-            position: Some(position),
-        }
-    }
-}
 
 /// Modify the positions of the channels.
 ///
 /// The minimum amount of channels to modify, is a swap between two channels.
-///
-/// This function accepts an `Iterator` of `(Id<ChannelMarker>, u64)`. It also accepts
-/// an `Iterator` of `Position`, which has extra fields.
 #[must_use = "requests must be configured and executed"]
 pub struct UpdateGuildChannelPositions<'a> {
     guild_id: Id<GuildMarker>,
